utils: move dict_set
