+++ invenio_rdm_migrator/load/postgresql/generators.py
@@ -10,36 +10,7 @@
 from abc import ABC, abstractmethod
 
 from ...logging import Logger
-
-
-# PORT: from invenio_records.dictutils to avoid having an invenio constraint
-# it could cause troubles with sqlalchemy and psycompg version
-def parse_lookup_key(lookup_key):
-    """Parse a lookup key."""
-    if not lookup_key:
-        raise KeyError("No lookup key specified")
-
-    # Parse the list of keys
-    if isinstance(lookup_key, str):
-        keys = lookup_key.split(".")
-    elif isinstance(lookup_key, list):
-        keys = lookup_key
-    else:
-        raise TypeError("lookup must be string or list")
-
-    return keys
-
-
-def dict_set(source, key, value):
-    """Set a value into a dict via a dot-notated key."""
-    keys = parse_lookup_key(key)
-    parent = source
-    for key in keys[:-1]:
-        if isinstance(key, int):
-            parent = parent[key]
-        else:
-            parent = parent.setdefault(key, {})
-    parent[keys[-1]] = value
+from ...utils import dict_set
 
 
 class PostgreSQLGenerator(ABC):
