+++ invenio_rdm_migrator/utils.py
@@ -29,3 +29,33 @@ class JSONEncoder(json.JSONEncoder):
             return str(o)
         # Let the base class default method raise the TypeError
         return json.JSONEncoder.default(self, o)
+
+
+# PORT: from invenio_records.dictutils to avoid having an invenio constraint
+# it could cause troubles with sqlalchemy and psycompg version
+def parse_lookup_key(lookup_key):
+    """Parse a lookup key."""
+    if not lookup_key:
+        raise KeyError("No lookup key specified")
+
+    # Parse the list of keys
+    if isinstance(lookup_key, str):
+        keys = lookup_key.split(".")
+    elif isinstance(lookup_key, list):
+        keys = lookup_key
+    else:
+        raise TypeError("lookup must be string or list")
+
+    return keys
+
+
+def dict_set(source, key, value):
+    """Set a value into a dict via a dot-notated key."""
+    keys = parse_lookup_key(key)
+    parent = source
+    for key in keys[:-1]:
+        if isinstance(key, int):
+            parent = parent[key]
+        else:
+            parent = parent.setdefault(key, {})
+    parent[keys[-1]] = value
